feat(libvolcano): Create skeleton libretro-core
There are some function in this libretro-implementation that are currently
unused.

Since we still have to write out the paremeter-list to match the interface
but dont use any of them we should declare the parameters as unused.

That's why we use the C++17 attribute `[[maybe_unused]]` since it's standard
it should be actually portable unlike other solutions.

Also empty funtion bodies are now just two braces on the same line because
it looks neater :)