feat(volcano, renderer): Create a depth-buffer
Adds code to create a depth-buffer image in device-memory and attach it
to the framebuffers.

Also extends the renderpass to have a depth-buffer attachment slot.

Depth-tests are enables in the pipeline now as well.

Finally, this adds clearing of the depth-buffer to the dispatch-method
so we can draw more than one frame.