feat(volcano, renderer): Create an abstraction for cameras
This is just a simple container that can be used to keep track of
view- and projection-matrices for a lookat-style camera.

This class also contains a simple mechanism for lazy updates of
the resulting world->clip-space matrix:
  It tracks if it's high-level paramters have been changes since
  the last update with a dirty-flag and allows the user to trigger
  a recalculation of the matrix when it is convenient.

The idea is to avoid recalculating the final camera-matrix
everytime a user changes - for example - the eye-position even
though that result might never be used because of a later parameter
change.