feat(kepler): Create new class for handling transforms
This introduces two new classes for storing transforms and creating
references to them in a decoupled manner.

The `transform`-class has a dirty-flag and version-attribute to track
modifications of the transform.

Additionally the `transform_reference`-class can be used to obtain
a reference to such a transform, that keeps track of updates to
the referenced transform.

This allows lazy recalculation of the transform's matrix to save some
unnecessary CPU cycles.