feat(libvolcano): Check, if we are actually moving
Due to the way, we are tracking keypresses/movement it's possible to register
a move-event with a movement-vector that has zero magnitude (it's caused by
pressing both keys for a direction, giving a sum of zero for that axis).

Since we are calling `glm::normalize()` on our movement-vector to make movement
in all directions have equal magnitude we have to check for zero-magnitude
because otherwise we'll divide by zero.